make component and screen file
