+++ src/components/Maps.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import {View, StyleSheet} from 'react-native';
+import MapView from 'react-native-maps';
+
+const Maps = () => {
+  return (
+    <View style={{flex: 1}}>
+      <MapView
+        style={styles.map}
+        initialRegion={{
+          latitude: 37.78825,
+          longitude: -122.4324,
+          latitudeDelta: 0.0922,
+          longitudeDelta: 0.0421,
+        }}
+      />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  map: {
+    ...StyleSheet.absoluteFillObject,
+  },
+});
+
+export default Maps;
