+++ src/screens/MapScreen.tsx
@@ -1,0 +1,13 @@
+import React, {Component} from 'react';
+import {Text, StyleSheet, View} from 'react-native';
+import Maps from '../components/Maps';
+
+export const MapScreen = () => {
+  return (
+    <View>
+      <Maps />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({});
